+++ src/systems/phone/esp_ui_phone_manager.hpp
@@ -6,9 +6,11 @@
 #pragma once
 
 #include <memory>
+#include "lvgl.h"
 #include "core/esp_ui_core_manager.hpp"
 #include "widgets/gesture/esp_ui_gesture.hpp"
 #include "esp_ui_phone_home.hpp"
+#include "esp_ui_phone_app.hpp"
 
 class ESP_UI_Phone;
 
@@ -30,21 +32,25 @@ protected:
     const ESP_UI_PhoneManagerData_t &_data;
 
 private:
-    // Main
-    bool begin(void);
-    bool del(void);
+    // Core
     bool processAppRunExtra(ESP_UI_CoreApp *app) override;
     bool processAppResumeExtra(ESP_UI_CoreApp *app) override;
     bool processAppCloseExtra(ESP_UI_CoreApp *app) override;
+    bool processNavigationEvent(ESP_UI_CoreNavigateType_t type) override;
+    // Main
+    bool begin(void);
+    bool del(void);
     // Home
+    bool processHomeScreenChange(ESP_UI_PhoneManagerScreen_t screen, void *data);
     static void onHomeMainScreenLoadEventCallback(lv_event_t *event);
     // App Launcher
     static void onAppLauncherGestureEventCallback(lv_event_t *event);
-    // Navigation
-    bool processNavigationEvent(ESP_UI_CoreNavigateType_t type) override;
-    static void onNavigationGesturePressingEventCallback(lv_event_t *event);
-    static void onNavigationGestureReleaseEventCallback(lv_event_t *event);
-    // RecentsScreen
+    // Navigation Bar
+    static void onNavigationBarGestureEventCallback(lv_event_t *event);
+    // Gesture
+    static void onGestureNavigationPressingEventCallback(lv_event_t *event);
+    static void onGestureNavigationReleaseEventCallback(lv_event_t *event);
+    // Recents Screen
     bool processRecentsScreenShow(void);
     bool processRecentsScreenHide(void);
     bool processRecentsScreenMoveLeft(void);
@@ -56,14 +62,20 @@ private:
 
     // Main
     bool _is_initialized;
+    ESP_UI_PhoneManagerScreen_t _home_active_screen;
     // App Launcher
-    // The following variables are used to prevent the app launcher from being operated when the app is closed
-    bool _is_app_closed;
     bool _is_app_launcher_gesture_disabled;
     ESP_UI_GestureDirection_t _app_launcher_gesture_dir;
-    // Navigation
-    bool _enable_navigation_gesture;
-    bool _is_navigation_gesture_released;
+    // Navigation Bar
+    bool _enable_navigation_bar_gesture;
+    bool _is_navigation_bar_gesture_disabled;
+    ESP_UI_GestureDirection_t _navigation_bar_gesture_dir;
+    // Gesture
+    bool _enable_gesture_navigation;
+    bool _enable_gesture_navigation_back;
+    bool _enable_gesture_navigation_home;
+    bool _enable_gesture_navigation_recents_app;
+    bool _is_gesture_navigation_disabled;
     std::unique_ptr<ESP_UI_Gesture> _gesture;
     // RecentsScreen
     bool _recents_screen_pressed;